requirements.txt
asyncssh>=2.14.0
aiohttp>=3.9.0
flask>=3.0.0
flask-cors>=4.0.0
cryptography>=41.0.0
python-dotenv>=1.0.0
markdown2>=2.4.0
weasyprint>=60.0
